[25.11.2024] Added color theme and refactored some files
