+++ rescue_now_app/lib/src/crash_detection.dart
