+++ rescue_now_app/lib/theme/app_theme.dart
@@ -1,0 +1,15 @@
+import 'package:flutter/material.dart';
+import 'package:rescue_now_app/theme/colors.dart';
+
+@immutable
+class AppTheme {
+  static const colors = AppColors();
+
+  const AppTheme._();
+
+  static ThemeData define() {
+    return ThemeData(
+      fontFamily: "Roboto",
+    );
+  }
+}
