+++ rescue_now_app/lib/theme/colors.dart
@@ -1,0 +1,13 @@
+import 'package:flutter/material.dart';
+
+
+// Global colors for the app
+@immutable
+class AppColors {
+  final background = const Color(0xFFF6DBD5);
+  final primary = const Color(0xFF885053);
+  final menuButtons = const Color(0xFF4C3527);
+  final text = const Color(0xFFFFFFFF);
+
+  const AppColors();
+}
