+++ rescue_now_app/test/widget_test.dart
@@ -1,6 +1,6 @@
 import 'package:flutter_test/flutter_test.dart';
 
-import 'package:rescue_now_app/crash_detection.dart'; // Update this to the correct path if necessary
+import 'package:rescue_now_app/src/crash_detection.dart'; // Update this to the correct path if necessary
 
 void main() {
   testWidgets('Crash Detection App Test', (WidgetTester tester) async {
